+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## v6.0.1 - 2023-10-20
 
 ### Changes
 - update @mia-platform/lc39 lib to v7.0.2 to fix the "fastify.close is not a function" error